% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_yaml.R
\name{check.yaml}
\alias{check.yaml}
\title{Static analysis of YAML metadata specification}
\usage{
check.yaml(file = NULL)
}
\arguments{
\item{file}{}
}
\description{
Static analysis of YAML metadata specification
}
