% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_metadata.R
\name{check_metadata}
\alias{check_metadata}
\title{Check \code{epicdata_meta} class objects}
\usage{
check_metadata(metadata)
}
\arguments{
\item{metadata}{An \code{epicdata_meta} object.}
}
\value{
Prints information about successful and unsuccessful checks in the
console. Returns the input metadata. If checks were successful, returns
checked metadata that can be used in data processing.
}
\description{
Check if \code{epicdata_meta} class objects fulfill the conditions for being used
in data processing.
}
\examples{
x <- empty_metadata()
check_metadata(x)
}
