% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_projects.R
\name{create.metadata.file}
\alias{create.metadata.file}
\title{Create new metadata file}
\usage{
create.metadata.file(name, path = getwd())
}
\arguments{
\item{name}{The name of the file without file extension.}

\item{path}{The path where the file should be saved, without the name of
the file. Defaults to the current working directory.}
}
\value{
Creates and opens a template .yml file for creating metadata.
}
\description{
Create a new YAML file for specifying metadata specifications.
}
\examples{
withr::with_tempdir({
create.metadata.file("my_data_set")
})
}
