% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_projects.R
\name{create_prc}
\alias{create_prc}
\title{Creating a prc project}
\usage{
create_prc(name, path, db = F)
}
\arguments{
\item{name}{The name of the project as a single character element. It is
recommended to only use lower case letters, numbers, as well as _ and to
end the name with '_prc'.}

\item{path}{The path where the project should be created, without the name of
the project.}

\item{db}{A TRUE/FALSE value indicating if the created project should include
code that supports connection to a data base via DSN in R Studio. The default
is FALSE.}
}
\value{
Creates an R project in the specified library and adds all templates
and structre needed for processing data.
}
\description{
Creates a project that processes the data
}
