% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset}
\alias{dataset}
\title{Dataset objects}
\usage{
dataset(raw, metadata)
}
\arguments{
\item{raw}{Raw data.frame, data.table, or tibble}

\item{metadata}{A S7 \code{epicdata::metadata} object}
}
\value{
A S7 \code{epicdata::dataset} object
}
\description{
S7 class \code{epicdata::metadata}
}
\section{Additional properties}{

\describe{
\item{\code{@processed}}{A data.frame, data.table, or tibble containing the current
state of processing of the \code{raw} data}

\item{\code{@decisions}}{A list of the decisions taken during data processing}

\item{\code{@log}}{A list logging all changes and other relevant information during
data processing}
}}

