% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicdata-package.R
\docType{package}
\name{epicdata-package}
\alias{epicdata}
\alias{epicdata-package}
\title{epicdata: Data Processing in Epidemiology}
\description{
A workflow for processing of epidemiological data, including setup of meta data and handling of raw data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/forsterepi/epicdata}
  \item Report bugs at \url{https://github.com/forsterepi/epicdata/issues}
}

}
\author{
\strong{Maintainer}: Felix Forster \email{felix.forster@med.uni-muenchen.de} (\href{https://orcid.org/0000-0002-3670-9244}{ORCID})

}
\keyword{internal}
