% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicdata_meta.R
\name{new_meta}
\alias{new_meta}
\alias{validate_meta}
\alias{print.epicdata_meta}
\title{Metadata objects}
\usage{
new_meta(x = list())

validate_meta(x)

\method{print}{epicdata_meta}(x, ...)
}
\arguments{
\item{x}{A list to be converted to class \code{epicdata_meta}.}

\item{...}{Additional arguments for generics \code{print()}, \code{summary()}, and \code{plot()}.}
}
\value{
A \code{epicdata_meta} object. Its structure is described in
Details.
}
\description{
Constructor function for S3 class \code{epicdata_meta}.

\code{new_meta()} and \code{validate_meta()} define the S3 class.

\code{print()} prints
}
\details{
\code{epicdata_meta} objects are lists containing 7 data.frames.
These data.frames are described below:

\code{options} contains

\code{var.list} contains

\code{na.codes} contains

\code{na.rules} contains

\code{contras} contains

\code{mc.sets} contains

\code{dict} contains
}
\examples{
x <- list(data.frame(a = c(1,2,3), b = c(4,5,6)),
data.frame(a = c("a","b","c"), b = c("d","e","f")))
metadata <- new_meta(x)
}
