% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_meta_creator.R
\name{launch_meta_creator}
\alias{launch_meta_creator}
\title{Launch meta creator \code{shiny} app}
\usage{
launch_meta_creator()
}
\value{
The \code{launch_meta_creator} function is used for the side effect of starting the Meta Data Creator \code{shiny} app.
}
\description{
Launch meta creator \code{shiny} app
}
\examples{
\dontrun{
launch_meta_creator()
}
}
