% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epic_metadata.R
\name{load_metadata}
\alias{load_metadata}
\title{Load \code{epic_metadata} objects}
\usage{
load_metadata(path)
}
\arguments{
\item{path}{File path to \code{epic_metadata} object.}
}
\value{
Load a \code{epic_metadata} object and returns it.
}
\description{
Load \code{epic_metadata} objects
}
\examples{
withr::with_tempdir({
x <- empty_metadata()
saveRDS(x,"metadata.rds")
metadata <- load_metadata("metadata.rds")})
}
