% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata.all.vars.of.type}
\alias{metadata.all.vars.of.type}
\title{Extract the names of all variables of a certain type}
\usage{
metadata.all.vars.of.type(metadata, type)
}
\arguments{
\item{metadata}{A S7 \code{epicdata::metadata} object}

\item{type}{A single string, one of \code{text}, \code{cat}, \code{num}, \code{date}, \code{time},
\code{datetime}}
}
\value{
A character string with the names of all variables of the provided
type in the order of appearance in the metadata.
}
\description{
Extract the names of all variables of a certain type
}
