% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\title{Metadata objects}
\usage{
metadata(file)
}
\arguments{
\item{file}{Path to yaml file}
}
\value{
A S7 \code{epicdata::metadata} object
}
\description{
S7 class \code{epicdata::metadata}
}
\section{Additional properties}{

\describe{
\item{\code{@data.name}}{The name of the data.frame. A single string.}

\item{\code{@id.var}}{The name of the ID variable. A variable name used in \code{var.list}.}
}}

