% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epic_metadata.R
\name{new_metadata}
\alias{new_metadata}
\title{Construct \code{epic_metadata} objects}
\usage{
new_metadata(x = list())
}
\arguments{
\item{x}{A list to be converted to class \code{epic_metadata}.}
}
\value{
A \code{epic_metadata} object. Its structure is described in
Details.
}
\description{
Constructor function for S3 class \code{epic_metadata}. The class is used
with the Meta Data Creator \code{shiny} app to store the inputed information.
}
\details{
\code{epic_metadata} objects are lists containing 6 data.frames.
These data.frames are described below:

\code{main} contains

\code{missing_codes} contains

\code{missing_rules} contains

\code{contras} contains

\code{mc_sets} contains

\code{string_dict} contains
}
\examples{
x <- list(data.frame(a = c(1,2,3), b = c(4,5,6)),
data.frame(a = c("a","b","c"), b = c("d","e","f")))
metadata <- new_metadata(x)
}
\seealso{
\code{\link[=validate_metadata]{validate_metadata()}}, \code{\link[=print.epic_metadata]{print.epic_metadata()}},
\code{\link[=empty_metadata]{empty_metadata()}}
}
