% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epic_metadata.R
\name{print.epic_metadata}
\alias{print.epic_metadata}
\title{Print \code{epic_metadata} objects}
\usage{
\method{print}{epic_metadata}(x, ...)
}
\arguments{
\item{x}{A \code{epic_metadata} object.}

\item{...}{further arguments to be passed to or from other
    methods.  They are ignored in this function.}
}
\value{
Prints the \code{epic_metadata} object in the console.
}
\description{
Print method for objects of class \code{epic_metadata}.
}
\examples{
x <- empty_metadata()
print(x)
}
\seealso{
\code{\link[=new_metadata]{new_metadata()}} for information on \code{epic_metadata} objects
}
