% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicdata_meta.R
\name{print.epicdata_meta_checked}
\alias{print.epicdata_meta_checked}
\title{Print \code{epicdata_meta_checked} objects}
\usage{
\method{print}{epicdata_meta_checked}(x, ...)
}
\arguments{
\item{x}{A \code{epicdata_meta_checked} object.}

\item{...}{further arguments to be passed to or from other
    methods.  They are ignored in this function.}
}
\value{
Prints the \code{epicdata_meta_checked} object in the console.
}
\description{
Print method for objects of class \code{epicdata_meta_checked}.
}
\examples{
x <- empty_metadata()
x <- check_metadata(x)
print(x)
}
\seealso{
\code{\link[=new_metadata]{new_metadata()}} for information on \code{epicdata_meta} objects
}
