% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_df}
\alias{sim_df}
\title{Simulated example data.frame}
\format{
A data.frame with 200 rows and 15 columns:
\describe{
\item{part_ids}{Participant ID with duplicates indicating multiply entered answers from a non-digital medium, e.g., a paper questionnaire.}
\item{male}{A binary variable indicating sex.}
\item{asthma_diag}{A categorical variable with 3 categories indicating presence of asthma diagnosis.}
\item{height}{A numerical variable indicating height in cm.}
\item{weight}{A numerical variable indicating weight in kg.}
\item{mc1}{The first item of a multiplce choice question.}
\item{mc2}{The second item of a multiplce choice question.}
\item{mc3}{The third item of a multiplce choice question.}
\item{mc4}{The fourth item of a multiplce choice question.}
\item{mc_other}{The 'other' option of a multiplce choice question.}
\item{mc_ops}{The 'other: please specify' text field of a multiple choice question.}
\item{jobs}{A character variable describing the participant's job.}
\item{birthday}{A date variable in format yyyy-mm-dd.}
\item{date_of_something}{A date variable in format (d)d.(m)m.yyyy.}
\item{entry_id}{An ID, which is unique for every entry, i.e., row. Similar to rownumbers.}
}
}
\usage{
sim_df
}
\description{
A data.frame, which has been simulated to support testing and examples. All variables are of type character.
}
\keyword{datasets}
