% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_of_data.R
\name{state.of.the.data}
\alias{state.of.the.data}
\title{Evaluate the state of the data}
\usage{
state.of.the.data(d, metadata, when, all.vars.character = NULL)
}
\arguments{
\item{d}{A data.frame, tibble, or data.table to be evaluated}

\item{metadata}{A S7 \code{epicdata::metadata} object}

\item{when}{A single string, either \code{before} or \code{after}, indicating if the
state of the data is evaluated at the start or end of a workflow step.
\code{before} will produce errors for users, while \code{after} will produce internal
errors for the developer.}

\item{all.vars.character}{TRUE or NULL (default). TRUE evaluates the
characteristic, while NULL skips evaluation. FALSE does not make sense as it
would result in incorrect errors for datasets that really consist only of
variables of type \code{text}. Thus, FALSE will be converted to NULL.}
}
\value{
The input dataset invisibly. However, the function is called for its
side effects.
}
\description{
Evaluates a set of characteristics of the dataset to be processed in order to
ensure that combining workflow steps by the user works as expeceted. The
state of the data is evaluated before and after a workflow step, i.e., at the
start and at the end of the corresponding function.
}
\details{
See workflow vignette for details
}
