% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicdata_meta.R
\name{validate_metadata}
\alias{validate_metadata}
\title{Validate \code{epicdata_meta} objects}
\usage{
validate_metadata(x)
}
\arguments{
\item{x}{An \code{epicdata_meta} or \code{epicdata_meta_checked} object to be
validated.}
}
\value{
An \code{epicdata_meta} or \code{epicdata_meta_checked} object, that
has been checked to have the correct structure.
}
\description{
Validates that the provided input object of class \code{epicdata_meta} has
the correct structure.
}
\examples{
x <- empty_metadata()
x <- validate_metadata(x)
}
\seealso{
\code{\link[=new_metadata]{new_metadata()}} for information on \code{epicdata_meta} objects
}
