% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_VAR.R
\name{VAR_DF.NOT.META_RM}
\alias{VAR_DF.NOT.META_RM}
\title{Remove variables that are in the data but not in the metadata}
\usage{
VAR_DF.NOT.META_RM(dataset)
}
\arguments{
\item{dataset}{A S7 \code{epicdata::dataset} object}
}
\value{
A S7 \code{epicdata::dataset} object
}
\description{
Variant ID: VAR_DF.NOT.META_RM
}
