% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{variable}
\alias{variable}
\title{Variable objects}
\usage{
variable(
  name = stop("@name is required"),
  type = stop("@type is required"),
  old.id = NULL,
  label = NULL,
  label.eng = NULL,
  cats = NULL,
  cats.eng = NULL,
  touch.na = NULL,
  touch.na.options.default = NULL
)
}
\arguments{
\item{name}{The name of the variable}

\item{type}{x}

\item{old.id}{x}

\item{label}{x}

\item{label.eng}{x}

\item{cats}{x}

\item{cats.eng}{x}
}
\value{
An S7 \code{epicdata::variable} object
}
\description{
S7 class \code{epicmodel::variable}
}
