% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_projects.R
\name{create_meta}
\alias{create_meta}
\title{Creating a meta project}
\usage{
create_meta(name, path)
}
\arguments{
\item{name}{The name of the project as a single character element. It is
recommended to only use lower case letters, numbers, as well as _ and to
end the name with '_meta'.}

\item{path}{The path where the project should be created, without the name of
the project.}
}
\value{
Creates a R project in the specified library and adds all templates
and structre needed for creating meta data.
}
\description{
Creates a project that sets up meta data
}
