% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epic_metadata.R
\name{empty_metadata}
\alias{empty_metadata}
\title{Create empty \code{epic_metadata} object}
\usage{
empty_metadata()
}
\value{
A empty \code{epic_metadata} object with correct structure.
}
\description{
Creates an empty \code{epic_metadata} object with the correct structure. It
is used when starting up the Meta Data Creator \code{shiny} app.
}
\examples{
x <- empty_metadata()
}
\seealso{
\code{\link[=new_metadata]{new_metadata()}} for information on \code{epic_metadata} objects
}
