% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicdata-package.R
\docType{package}
\name{epicdata-package}
\alias{epicdata}
\alias{epicdata-package}
\title{epicdata: Clean Data for Epidemiology}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A workflow for processing of epidemiological data, turning your raw data into data that can be analyzed. You provide raw data and specify metadata in YAML using a user-friendly format. Decisions regarding invalid values can be collected in spreadsheet files. `epicdata` is thus suitable for non-veteran R users.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/forsterepi/epicdata}
  \item Report bugs at \url{https://github.com/forsterepi/epicdata/issues}
}

}
\author{
\strong{Maintainer}: Felix Forster \email{felix.forster@med.uni-muenchen.de} (\href{https://orcid.org/0000-0002-3670-9244}{ORCID}) [copyright holder]

}
\keyword{internal}
