% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\title{Metadata objects}
\usage{
metadata(file)
}
\arguments{
\item{file}{Path to yaml file}
}
\value{
An S7 \code{epicdata::metadata} object
}
\description{
S7 class \code{epicmodel::metadata}
}
