% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\title{Metadata objects}
\usage{
metadata(file)
}
\arguments{
\item{file}{Path to yaml}
}
\value{
A \code{epicdata_meta} object. Its structure is described in
Details.
}
\description{
Constructor function for S3 class \code{epicdata_meta}.
}
\details{
\code{epicdata_meta} objects are lists containing 5 data.frames.
These data.frames are described below:

\code{options} contains

\code{var.list} contains

\code{na.codes} contains

\code{na.rules} contains

\code{contras} contains
}
